chore(package): move some astParser related to file system into core
