+++ packages/babel-lit-jsx-plugin/package.json
@@ -18,7 +18,7 @@
     "url": "git+https://github.com/Shinyaigeek/lit-jsx.git"
   },
   "scripts": {
-    "test": "echo \"Error: run tests from root\" && exit 1"
+    "build": "tsc src/main.ts"
   },
   "bugs": {
     "url": "https://github.com/Shinyaigeek/lit-jsx/issues"
