+++ packages/babel-lit-jsx-plugin/src/main.ts
@@ -1,0 +1,7 @@
+import { visitor } from "core";
+
+export default function () {
+  return {
+    visitor,
+  };
+}
