+++ packages/core/visitor/convertTsx2TemplateLiteral/convertTsx2TemplateLiteral.tsx
@@ -1,8 +1,6 @@
 import traverse from "@babel/traverse";
 import { File } from "@babel/types";
 import { join } from "path";
-// import { resolvePath } from "../../file/resolvePath/resolvePath";
-import { compileEachFile } from "../compileEachFile/compileEachFile";
 import { convertComponent2Function } from "./convertComponent2Function/convertComponent2Function";
 import { convertReturnedJSXElementToString } from "./convertReturnedJSXElementToString/convertReturnedJSXElementToString";
 import { justifyJSXProps } from "./justifyJSXProps/justifyJSXProps";
