+++ packages/jsx2templateLiteral/src/modules/astParser/astParser.ts
@@ -1,6 +1,6 @@
 import { parse } from "@babel/parser";
 import { join } from "path";
-import { fileReader } from "../../file/fileReader/fileReader";
+import { fileReader } from "../file/fileReader/fileReader";
 
 export const astParser = (entry: string) => {
   const source = fileReader(entry);
